arthas cli工程完善

+++ kotlin-debugger-project/kotlin-debugger-arthas/kotlin-debugger-arthas-client/src/main/kotlin/com/wanna/middleware/arthas/core/config/Configure.kt
@@ -7,7 +7,7 @@ package com.wanna.middleware.arthas.core.config
  * @date 2023/2/25
  */
 class Configure {
-    var javaPid: Int = -1
+    var javaPid: Long = -1
 
     var arthasAgent: String? = null
